Split ReferenceDataContent out of Data
